save all

+++ runner/audio.js
@@ -1,19 +1,30 @@
 const { spawn } = require('child_process');
 
-module.exports.play = function(path, volume, callback) {
+module.exports.play = function(path, opts, callback) {
 
   console.log(`audio: playing ${path}`)
 
-  const process = spawn('play', ['-v', volume, path], {stdio: 'ignore'})
+  const process = spawn('play', ['-v', opts.volume, path].concat(opts.extraArgs), {stdio: 'ignore'})
 
   process.on('exit', code => {
-    console.log(`aplay finished with code ${code}`)
+    console.log(`play finished with code ${code}`)
     callback()
   })
 
-  return () => {
-    console.log(`audio : killing ${path}`)
-    process.kill('SIGKILL')
+  let killed = false;
+  const kill = () => {
+    if (!killed) {
+      console.log(`audio : killing ${path}`)
+      process.kill('SIGKILL')
+      killed = true;
+    }
   }
-}
 
+  if (opts.timeout != 0){
+    setTimeout(() => {
+      kill()
+    }, opts.timeout * 1000);
+  }
+
+  return kill;
+}
